requirements.txt
dnspython==2.4.0
asgiref==3.8.1
autocommand==2.2.2
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
djongo
django
dnspython==2.4.0
et_xmlfile==2.0.0
Flask==3.1.0
gunicorn
h11==0.14.0
httpcore==1.0.7
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
Markdown==3.7
MarkupSafe==3.0.2
mongo-connector==3.1.1
Mongo-db-operations==0.0.1
numpy==2.1.3
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
psycopg2-binary
pymongo==3.11.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sqlparse
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.1.3
zipp==3.21.0